Refer VSS integration proposal to IFEX docs
This is primarily a description about how to technically combine a
VSS data description and an IFEX (previously VSC) interface description.

It is not a discussion of the content of the VSS _standard catalog_
content in relation to the content of VSC catalog, which is what this
repo is about.

The description was also tailored towards the IFEX core IDL description
format.  Therefore, the now updated version stored in IFEX docs is more
appropriate to read and this file should be replaced to avoid confusion.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>